requirements.txt
flask
flask-cors
chromadb
numpy
python-dotenv
pytest
gunicorn
ollama